requirements.txt
alembic==1.4.2
alive-progress==1.5.1
arrow==0.15.7
b2sdk==1.1.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataset==1.3.2
dateutils==0.6.8
Deprecated==1.2.10
feedgen==0.9.0
Flask==1.1.2
funcsigs==1.0.2
future==0.18.2
humanize==2.5.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
logfury==0.1.2
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
mutagen==1.44.0
parse==1.15.0
podgen==1.1.0
PyGithub==1.51
PyJWT==1.7.1
pytaglib==1.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
stuf==0.9.16
tinys3==0.1.12
tinytag==1.4.0
tqdm==4.47.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
